+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_futures.dart
@@ -166,7 +166,7 @@ class FuturesAdapter extends OfflineFirstAdapter<Futures> {
   };
   Future<int> primaryKeyByUniqueColumns(
           Futures instance, DatabaseExecutor executor) async =>
-      null;
+      instance?.primaryKey;
   final String tableName = 'Futures';
   Future<void> afterSave(instance, {provider, repository}) async {
     if (instance.primaryKey != null) {