+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_offline_first_where.dart
@@ -167,7 +167,7 @@ class OfflineFirstWhereAdapter extends OfflineFirstAdapter<OfflineFirstWhere> {
   };
   Future<int> primaryKeyByUniqueColumns(
           OfflineFirstWhere instance, DatabaseExecutor executor) async =>
-      null;
+      instance?.primaryKey;
   final String tableName = 'OfflineFirstWhere';
   Future<void> afterSave(instance, {provider, repository}) async {
     if (instance.primaryKey != null) {