+++ packages/brick_offline_first_with_rest_build/test/offline_first_generator/test_one_to_many_association.dart
@@ -98,7 +98,7 @@ class OneToManyAssociationAdapter
   };
   Future<int> primaryKeyByUniqueColumns(
           OneToManyAssociation instance, DatabaseExecutor executor) async =>
-      null;
+      instance?.primaryKey;
   final String tableName = 'OneToManyAssociation';
   Future<void> afterSave(instance, {provider, repository}) async {
     if (instance.primaryKey != null) {