+++ packages/brick_sqlite/lib/src/sqlite_model_dictionary.dart
@@ -1,0 +1,8 @@
+import 'package:brick_core/core.dart';
+import 'package:brick_sqlite/src/sqlite_adapter.dart';
+import 'package:brick_sqlite_abstract/sqlite_model.dart';
+
+/// Associates app models with their [SqliteAdapter]
+class SqliteModelDictionary extends ModelDictionary<SqliteModel, SqliteAdapter<SqliteModel>> {
+  const SqliteModelDictionary(Map<Type, SqliteAdapter<SqliteModel>> mappings) : super(mappings);
+}