+++ packages/brick_sqlite/test/sqlite_test.dart
@@ -8,6 +8,7 @@ import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 import '__mocks__.dart';
+import '__mocks__/demo_model_adapter.dart';
 
 void main() {
   ft.TestWidgetsFlutterBinding.ensureInitialized();
@@ -20,11 +21,15 @@ void main() {
       modelDictionary: dictionary,
     );
 
-    setUpAll(() async {
+    setUp(() async {
       await provider.migrate([DemoModelMigration()]);
-      await provider.upsert<DemoModel>(DemoModel('Thomas'));
-      await provider.upsert<DemoModel>(DemoModel('Guy'));
-      await provider.upsert<DemoModel>(DemoModel('John'));
+      await provider.upsert<DemoModel>(DemoModel(name: 'ThomasDefault'));
+      await provider.upsert<DemoModel>(DemoModel(name: 'GuyDefault'));
+      await provider.upsert<DemoModel>(DemoModel(name: 'AliceDefault'));
+    });
+
+    tearDown(() async {
+      await provider.resetDb();
     });
 
     test('#get', () async {
@@ -34,14 +39,14 @@ void main() {
 
     group('#upsert', () {
       test('insert', () async {
-        final newModel = DemoModel('John');
+        final newModel = DemoModel(name: 'John');
 
         final newPrimaryKey = await provider.upsert<DemoModel>(newModel);
         expect(newPrimaryKey, 4);
       });
 
       test('update', () async {
-        final newModel = DemoModel('Guy')..primaryKey = 2;
+        final newModel = DemoModel(name: 'Guy')..primaryKey = 2;
 
         final primaryKey = await provider.upsert<DemoModel>(newModel);
         expect(primaryKey, 2);
@@ -85,15 +90,16 @@ void main() {
 
     group('#exists', () {
       test('specific', () async {
-        final newModel = DemoModel('John');
+        final newModel = DemoModel(name: 'Guy');
 
         await provider.upsert<DemoModel>(newModel);
-        final doesExist = await provider.exists<DemoModel>(query: Query.where('name', 'John'));
+        final doesExist =
+            await provider.exists<DemoModel>(query: Query.where('name', newModel.name));
         expect(doesExist, isTrue);
       });
 
       test('general', () async {
-        final newModel = DemoModel('John');
+        final newModel = DemoModel(name: 'Guy');
 
         await provider.upsert<DemoModel>(newModel);
         final doesExist = await provider.exists<DemoModel>();
@@ -106,14 +112,14 @@ void main() {
       });
 
       test('with an offset', () async {
-        await provider.upsert<DemoModel>(DemoModel('John'));
+        await provider.upsert<DemoModel>(DemoModel(name: 'Guy'));
         final existingModels = await provider.get<DemoModel>();
         final query = Query(providerArgs: {'limit': 1, 'offset': existingModels.length});
 
         final doesExistWithoutModel = await provider.exists<DemoModel>(query: query);
         expect(doesExistWithoutModel, isFalse);
 
-        await provider.upsert<DemoModel>(DemoModel('John'));
+        await provider.upsert<DemoModel>(DemoModel(name: 'Guy'));
         final doesExistWithModel = await provider.exists<DemoModel>(query: query);
         expect(doesExistWithModel, isTrue);
       });
@@ -127,7 +133,7 @@ void main() {
         InsertTable(localTableName),
         InsertColumn(columnName, Column.varchar, onTable: localTableName),
         InsertTable(foreignTableName),
-        InsertColumn('name', Column.varchar, onTable: foreignTableName),
+        InsertColumn('full_name', Column.varchar, onTable: foreignTableName),
       ];
       final joinsTableName =
           InsertForeignKey.joinsTableName(columnName, localTableName: localTableName);
@@ -156,8 +162,8 @@ void main() {
         for (var command in oldTable) {
           await provider.rawExecute(command.statement);
         }
-        await provider.rawInsert('INSERT INTO `$foreignTableName` (name) VALUES ("Bowler")');
-        await provider.rawInsert('INSERT INTO `$foreignTableName` (name) VALUES ("Big")');
+        await provider.rawInsert('INSERT INTO `$foreignTableName` (full_name) VALUES ("Bowler")');
+        await provider.rawInsert('INSERT INTO `$foreignTableName` (full_name) VALUES ("Big")');
         await provider.rawInsert(
             'INSERT OR IGNORE INTO `$localTableName` ($columnName) VALUES (?)', ['[1,2,3]']);
         for (var command in table) {
@@ -176,7 +182,7 @@ void main() {
 
   group('SqliteModel', () {
     test('#primaryKey', () {
-      final m = DemoModel('Thomas');
+      final m = DemoModel(name: 'Thomas');
       expect(m.primaryKey, NEW_RECORD_ID);
 
       m.primaryKey = 2;
@@ -184,7 +190,7 @@ void main() {
     });
 
     test('#isNewRecord', () {
-      final m = DemoModel('Thomas');
+      final m = DemoModel(name: 'Thomas');
       expect(m.isNewRecord, isTrue);
     });
 
@@ -196,7 +202,7 @@ void main() {
     final a = DemoModelAdapter();
 
     test('#tableName', () {
-      expect(a.tableName, sqliteTableName);
+      expect(a.tableName, 'DemoModel');
     });
   });
 }