+++ packages/brick_sqlite_generators/test/sqlite_model_serdes_generator/test_field_with_type_argument.dart
@@ -29,22 +29,22 @@ class FieldWithTypeArgumentAdapter
   FieldWithTypeArgumentAdapter();
 
   final Map<String, Map<String, dynamic>> fieldsToSqliteColumns = {
-    'primaryKey': {
-      'name': '_brick_id',
-      'type': int,
-      'iterable': false,
-      'association': false,
-    },
-    'someField': {
-      'name': 'some_field',
-      'type': Map,
-      'iterable': false,
-      'association': false,
-    }
+    'primaryKey': RuntimeSqliteColumnDefinition(
+      association: false,
+      columnName: '_brick_id',
+      iterable: false,
+      type: int,
+    ),
+    'someField': RuntimeSqliteColumnDefinition(
+      association: false,
+      columnName: 'some_field',
+      iterable: false,
+      type: Map,
+    ),
   };
   Future<int> primaryKeyByUniqueColumns(
           FieldWithTypeArgument instance, DatabaseExecutor executor) async =>
-      null;
+      instance?.primaryKey;
   final String tableName = 'FieldWithTypeArgument';
 
   Future<FieldWithTypeArgument> fromSqlite(Map<String, dynamic> input,